# File-Feature-Sync/lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Based Daily Speaking Coach API with Gamification
 * OpenAPI spec version: 0.2.0
 */

export interface AuthResponse {
  token: string;
  userId: string;
  username: string;
  email: string;
  role: string;
  purpose: string;
}